Update port to 0.10.6 libssh:   - Update the port to use commit 10e09e273f69e149389b3e0e5d44b8c221c2e7f6     from libssh, i.e. 0.10.6 stable.   - Update tested version list.   - Update README with latest port information.   - Bump revision to 4.1.0.
Upcoming ESP32 Arduino 3.0.0 preparation.
  - Fix FTBFS for ESP32 and S3 with gcc 12.2.
  - Example WiFi and IP event loops updated using IDF 5 calls.
  - Bump major version to 4 since ESP32 Arduino 1.0.6 can no longer
    be supported with the new event loops.

+++ src/pki_container_openssh.c
@@ -172,10 +172,10 @@ static int pki_private_key_decrypt(ssh_string blob,
     }
 
     SSH_LOG(SSH_LOG_DEBUG,
-            "Decryption: %d key, %d IV, %d rounds, %zu bytes salt",
+            "Decryption: %d key, %d IV, %ld rounds, %zu bytes salt",
             cipher.keysize/8,
             cipher.blocksize,
-            rounds,
+            (long)rounds,
             ssh_string_len(salt));
 
     if (passphrase == NULL) {
@@ -299,12 +299,12 @@ ssh_pki_openssh_import(const char *text_key,
         goto out;
     }
     SSH_LOG(SSH_LOG_INFO,
-            "Opening OpenSSH private key: ciphername: %s, kdf: %s, nkeys: %d",
+            "Opening OpenSSH private key: ciphername: %s, kdf: %s, nkeys: %ld",
             ciphername,
             kdfname,
-            nkeys);
+            (long)nkeys);
     if (nkeys != 1) {
-        SSH_LOG(SSH_LOG_WARN, "Opening OpenSSH private key: only 1 key supported (%d available)", nkeys);
+        SSH_LOG(SSH_LOG_WARN, "Opening OpenSSH private key: only 1 key supported (%ld available)", (long)nkeys);
         goto out;
     }
 
@@ -478,9 +478,10 @@ static int pki_private_key_encrypt(ssh_buffer privkey_buffer,
         return SSH_ERROR;
     }
 
-    SSH_LOG(SSH_LOG_WARN, "Encryption: %d key, %d IV, %d rounds, %zu bytes salt",
+    SSH_LOG(SSH_LOG_WARN,
+                "Encryption: %d key, %d IV, %ld rounds, %zu bytes salt",
                 cipher.keysize/8,
-                cipher.blocksize, rounds, ssh_string_len(salt));
+                cipher.blocksize, (long)rounds, ssh_string_len(salt));
 
     if (passphrase == NULL){
         if (auth_fn == NULL){
